ci: explicitly specify token for codecov
Explicitly specify the codecov token to be used (i.e., `CODECOV_TOKEN`),
given that the latest v4 release of the codecov action requires it to be
able to generate coverage reports.
Additionally, fail CI if coverage reporting fails, since coverage is an
important enough metric for us to ensure that it's tracked consistently.